Implementation of UI functional Health Bar, with functional potion system and Damage implementation in Player and Enemy
Now we have a functional Health Bar that when the player takes damage, it reduces the number of hearts in the screen. System of potions for healing made with a cooldown system. Damage to the player, invincibility frames and damage to enemies implemented. There are still a couple of bugs with the damage collision with the 2 colliders in the main player.